[ADD] Index updated when same command is called
